+++ src/main/java/com/mbakara/service/CarService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface CarService {
 
     List<Car> getAllCars();
-    Car getCarById(int id);
+    Car getCarById(long id);
 
     void addCar(Car car);
     void removeCar(Car car);
